# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Baraka API",
        "description": "API backend pour la plateforme Baraka de microcrédit",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/baraka-backend/public",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/v1/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tableau de bord de l'administration",
                "operationId": "ce3e98c0785904b9a48bc90280eceb7c",
                "responses": {
                    "200": {
                        "description": "Données du tableau de bord admin"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/moderation/offres": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Liste paginée des offres en attente de validation",
                "operationId": "37cced690c17ce680e935b0449b179b3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des offres en_validation"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/moderation/offres/{id}/approuver": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approuver une offre (passe en statut actif)",
                "operationId": "573c9b9ead5afcbed2a7d75690a3d22e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre approuvée"
                    },
                    "404": {
                        "description": "Offre introuvable"
                    },
                    "422": {
                        "description": "L'offre n'est pas en attente de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/moderation/offres/{id}/rejeter": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Rejeter une offre (repasse en brouillon avec motif)",
                "operationId": "6135375e59791dbf61efc9ccbcd7bb00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motif"
                                ],
                                "properties": {
                                    "motif": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "minLength": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offre rejetée"
                    },
                    "404": {
                        "description": "Offre introuvable"
                    },
                    "422": {
                        "description": "L'offre n'est pas en attente de validation ou motif manquant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/imfs/{id}/suspendre": {
            "put": {
                "tags": [
                    "Admin - IMFs"
                ],
                "summary": "Suspendre une IMF et dépublier toutes ses offres actives",
                "operationId": "f07cec6fc6a9fe28873c3899b848889b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IMF suspendue"
                    },
                    "404": {
                        "description": "IMF introuvable"
                    },
                    "422": {
                        "description": "L'IMF est déjà suspendue"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/imfs/{id}/reactiver": {
            "put": {
                "tags": [
                    "Admin - IMFs"
                ],
                "summary": "Réactiver une IMF suspendue",
                "operationId": "9453d4aadaee414ed1d1bb4b709b9de7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IMF réactivée"
                    },
                    "404": {
                        "description": "IMF introuvable"
                    },
                    "422": {
                        "description": "L'IMF est déjà active"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/imfs/{id}": {
            "delete": {
                "tags": [
                    "Admin - IMFs"
                ],
                "summary": "Supprimer (soft delete) une IMF — super_admin uniquement",
                "operationId": "4a8a26677a661be424f5a7c297428217",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IMF supprimée"
                    },
                    "403": {
                        "description": "Réservé au super_admin"
                    },
                    "404": {
                        "description": "IMF introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/journal": {
            "get": {
                "tags": [
                    "Admin - Journal"
                ],
                "summary": "Liste paginée du journal d'administration — super_admin uniquement",
                "operationId": "29294efba1b985e984d20ddaba46fb5d",
                "parameters": [
                    {
                        "name": "id_admin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Journal des actions admin"
                    },
                    "403": {
                        "description": "Réservé au super_admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le token JWT obtenu via POST /api/auth/otp/verify"
            }
        }
    },
...
}